+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "@/App";
-import * as serviceWorker from "@/serviceWorker";
-import "@/assets/scss/main.scss";
+import App from "App";
+import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
+import "assets/scss/main.scss";
 import Axios from "axios";
 
 Axios.defaults.baseURL = process.env.REACT_APP_API_BASE_URL;
@@ -14,7 +14,11 @@ ReactDOM.render(
   document.getElementById("root")
 );
 
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
+serviceWorkerRegistration.register({
+  onUpdate: (registration) => {
+    if (registration.waiting) {
+      registration.waiting?.postMessage({ type: "SKIP_WAITING" });
+      window.location.reload();
+    }
+  },
+});
